Task3/event_store_function/requirements.txt
google-cloud-storage>=2.9.0